ops_dashboard/requirements.txt
altair<5.0.0
db-dtypes
folium
google-auth
google-cloud-bigquery<4.0.0
grpcio<2.0.0
gspread
numpy
pandas
plotly
protobuf>=3.20.0,<4.0.0
shapely
streamlit>=1.30.0
streamlit-folium
openpyxl
